A not so nice solution to the float problem
Convert it to a double and use two almost identical functions...

+++ co2_sensor/lib/generic_esp_32/generic_esp_32.c
@@ -206,7 +206,7 @@ int variable_sprintf_size(char *string, int count, ...)
             extraSize += snprintf(snBuf, 0, "%s", va_arg(list, char *));
             break;
         case 'f':
-            extraSize += snprintf(snBuf, 0, "%f", va_arg(list, float));
+            extraSize += snprintf(snBuf, 0, "%f", va_arg(list, double));
             break;
         }
     }
